fix(main): add float step to Number

+++ src/main/java/cc/polyfrost/oneconfig/config/annotations/Number.java
@@ -44,7 +44,7 @@ public @interface Number {
 
     float max();
 
-    int step() default 1;
+    float step() default 1f;
 
     int size() default 1;
